Reduce Large Files Step 4: extract task-runner phases
## Summary

This slice divides `task-runner.ts` into execution phase modules: prepare, dispatch, and finalize.

The phase host wires the modules together and keeps executor selection and callback semantics the same.

The orchestrator integration is unchanged.

<details>
<summary>Review metadata</summary>

Review Claim: Approve dividing the task runner into prepare, dispatch, and finalize phase modules with the same execution behavior.

Review Lane: refactor

Review Unit: routing

Safety Invariant: Phase modules preserve the existing callback contract and executor selection, so the task lifecycle behaves identically.

Slice Rationale: Phase boundaries isolate execution branching from setup and reporting, so each module reviews on its own.

</details>

## Non-goals

- No behavior change to task execution; behavior unchanged.
- No change to headless commands or persistence modules.

## Test Plan

- [ ] `cd packages/execution-engine && pnpm test`

## Revert Plan

- Safe to revert? Yes
- Revert command: `git revert <sha>`
- Post-revert steps: None
- Data migration? No
